Package the master branch, so we enable the AAF plugin
This is the AVTP Audio Format plugin, needed for Audio and Video over
network (Ethernet). The 1.1.8 release should be coming in early 2019,
but we can start providing new functionality before that.

The plugin is compiled by default but won't kick in unless the user
specifically selects it as a backend via the ALSA API (like the -D
option to aplay).

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>